mv GroupCoordinatorPatch to worker
